+++ tests/run-tutorial/tutorial.hpp
@@ -86,7 +86,7 @@ void Function9(double arg);
 void Function10();
 void Function10(const std::string &name, double arg2);
 
-void Sum(int len, int * values, int *result);
+void Sum(size_t len, int * values, int *result);
 
 long long TypeLongLong(long long arg1);
